requirements.txt
absl-py==0.11.0
anyio==2.2.0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bcrypt==3.1.7
betterproto==1.2.5
black==20.8b1
bleach==3.3.0
blinker==1.4
Brlapi==0.7.0
cachetools==4.2.1
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.3
colorcet==2.0.6
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
en==0.0.1
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
grpclib==0.4.1
h2==4.0.0
h5py==2.10.0
hpack==4.0.0
httplib2==0.14.0
hyperframe==6.0.0
idna==2.8
imageio==2.9.0
ipycanvas==0.8.2
ipydatawidgets==4.2.0
ipyevents==0.8.2
ipykernel==5.5.0
ipympl==0.5.8
ipython==7.21.0
ipython-genutils==0.2.0
ipyvtk-simple==0.1.4
ipywidgets==7.6.3
itk-core==5.1.2
itk-filtering==5.1.2
itk-meshtopolydata==0.6.3
itk-numerics==5.1.2
itkwidgets==0.32.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.4.1
jupyterlab==3.0.12
jupyterlab-pygments==0.1.2
jupyterlab-server==2.3.0
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.3
MarkupSafe==1.1.0
matplotlib==3.3.4
meshio==4.3.10
mistune==0.8.4
monotonic==1.5
multidict==5.1.0
mypy-extensions==0.4.3
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nbtlib==1.12.0
nest-asyncio==1.5.1
netifaces==0.10.4
notebook==6.2.0
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
orjson==3.5.1
packaging==20.9
pandocfilters==1.4.3
param==1.10.1
paramiko==2.6.0
parso==0.8.1
pathspec==0.8.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pkg-resources==0.0.0
prometheus-client==0.9.0
prompt-toolkit==3.0.17
protobuf==3.14.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pyct==0.4.8
pycups==1.9.73
Pygments==2.8.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
PyQt5==5.15.3
PyQt5-Qt==5.15.2
PyQt5-sip==12.8.1
pyRFC3339==1.1
pyrsistent==0.17.3
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyvista==0.28.1
pyvistaqt==0.3.0
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.0.3
QtPy==1.9.0
regex==2020.11.13
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.1
sacremoses==0.0.43
scooby==0.5.6
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.16.0
six==1.15.0
sniffio==1.2.0
stringcase==1.2.0
systemd-python==234
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.3
testpath==0.4.4
tokenizers==0.10.1
toml==0.10.2
torch==1.7.1+cpu
torchaudio==0.7.2
torchvision==0.8.2+cpu
tornado==6.1
tqdm==4.57.0
traitlets==5.0.5
traittypes==0.2.1
transformers==4.3.2
transforms3d==0.3.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vtk==9.0.1
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xkit==0.0.0
zstandard==0.15.2
